FileAPI/tests/submissions/Opera/Blob-slice.html

<!doctype html>
<html>
 <head>
  <title>Blob slice</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
 </head>
 <body>
  <div id="log"></div>
  <script>
    test(function(){
        var b = new Blob();
        var prefixes = ['op', 'moz', 'webkit', 'ms'];
        for (var i = 0; i < prefixes.length; ++i) {
            assert_false(prefixes[i]+'Slice' in b, "'"+prefixes[i]+"Slice' in b");
            assert_false(prefixes[i]+'Slice' in Blob.prototype, "'"+prefixes[i]+"Slice in Blob.prototype");
        }
        assert_true('slice' in b, "'slice' in b");
        assert_true('slice' in Blob.prototype, "'slice' in Blob.prototype");
    });
  </script>
 </body>
</html>
